+++ src/main/java/duke/AddDeadlineCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.io.IOException;
 
 public class AddDeadlineCommand extends Command {
