+++ src/main/java/duke/DeleteCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.util.List;
 
 public class DeleteCommand extends Command {
