+++ src/main/java/duke/DoneCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.util.List;
 
 public class DoneCommand extends Command {
