+++ src/main/java/duke/Duke.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class Duke {
 
     private Storage storage;
