+++ src/main/java/duke/GetCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.time.LocalDate;
 import java.util.List;
