+++ src/main/java/duke/ListCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.util.List;
 
 public class ListCommand extends Command {
